+++ plugins/example-todo-list-backend/src/service/router.ts
@@ -51,7 +51,7 @@ export async function createRouter(
 
   router.get('/health', (_, response) => {
     logger.info('PONG!');
-    response.send({ status: 'ok' });
+    response.json({ status: 'ok' });
   });
 
   router.get('/todos', async (_req, res) => {