+++ packages/backend-next/CHANGELOG.md
@@ -1,5 +1,15 @@
 # example-backend-next
 
+## 0.0.4-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-backend@1.7.0-next.1
+  - @backstage/backend-defaults@0.1.2-next.1
+  - @backstage/plugin-app-backend@0.3.37-next.1
+  - @backstage/plugin-catalog-backend@1.4.1-next.1
+
 ## 0.0.4-next.0
 
 ### Patch Changes