+++ plugins/airbrake-backend/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-airbrake-backend
 
+## 0.2.10-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.15.2-next.1
+  - @backstage/config@1.0.3-next.1
+
 ## 0.2.10-next.0
 
 ### Patch Changes