+++ plugins/catalog-graphql/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-catalog-graphql
 
+## 0.3.14-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.2-next.1
+  - @backstage/config@1.0.3-next.1
+  - @backstage/types@1.0.0
+
 ## 0.3.14-next.0
 
 ### Patch Changes