+++ plugins/rollbar-backend/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-rollbar-backend
 
+## 0.1.34-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.15.2-next.1
+  - @backstage/config@1.0.3-next.1
+
 ## 0.1.34-next.0
 
 ### Patch Changes