+++ plugins/todo-backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-todo-backend
 
+## 0.1.34-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-client@1.1.1-next.1
+  - @backstage/backend-common@0.15.2-next.1
+  - @backstage/catalog-model@1.1.2-next.1
+  - @backstage/config@1.0.3-next.1
+  - @backstage/errors@1.1.2-next.1
+  - @backstage/integration@1.3.2-next.1
+
 ## 0.1.34-next.0
 
 ### Patch Changes